file change: set up world and base files
created empty game object called world at 0,0,0 and parented the directional light and a terrain to it. created a player capsule placed it on the terrain at 20,1,20 tagged and named it Player. parented main camera to player at 0,1,0 to sit at head height for future, main camera is tagged MainCamera. Created Stats and Leveling folder created empty PlayerStats C# script which will hold stats and act as the space for saving and loading from Json in the future.